Fix parameter pickup when de/reactivating control
If state prior to deactivation had int/ext equal to each other, then on
reactivation the conditional logic for pickup would accept the new ext
value regardless of how far away it was from the internal.

Change logic to compare new value to internal and use difference
threshold to determine how to proceed.